+++ testing-and-versioning/.gitignore
@@ -1,0 +1,1 @@
+github.json
