--- testing-and-versioning/config.template.json
@@ -1,17 +1,0 @@
-{
-  "registry": {
-    "owner": "gmac",
-    "repo": "test-registry",
-    "authToken": "keepItSecretKeepItSafe"
-  },
-  "services": [
-    {
-      "name": "inventory",
-      "url": "http://localhost:4001",
-    },
-    {
-      "name": "products",
-      "url": "http://localhost:4002",
-    }
-  ]
-}
