+++ testing-and-versioning/lib/make_remote_executor.js
@@ -6,7 +6,7 @@ const AbortController = require('abort-controller');
 // customize any way that you need (auth, headers, etc).
 // Expects to recieve an object with "document" and "variable" params,
 // and asynchronously returns a JSON response from the remote.
-module.exports = function makeRemoteExecutorWithTimeout(url, { timeout=500, headers={} }={}) {
+module.exports = function makeRemoteExecutor(url, { timeout=500, headers={} }={}) {
   return async ({ document, variables }) => {
     const controller = new AbortController();
     const timeoutId = setTimeout(() => controller.abort(), timeout);
