--- testing-and-versioning/services/.gitignore
@@ -1,1 +1,0 @@
-config.json
